Move nofity goroutine to txpool
