+++ zk/txpool/pool.go
@@ -422,9 +422,23 @@ func New(newTxs chan types.Announcements, coreDB kv.RoDB, cfg txpoolcfg.Config, 
 	tp.setFreeGasList(ethCfg.DeprecatedTxPool.FreeGasList)
 
 	// For X Layer
-	if ethCfg.XLayer.BulkAddTxs && ethCfg.DeprecatedTxPool.EnableNotify {
-		notificationStreams = NewTxpoolNotificationStreams()
-	}
+	notifyOnce.Do(
+		func() {
+			if ethCfg.XLayer.BulkAddTxs && ethCfg.DeprecatedTxPool.EnableNotify {
+				needNofity = true
+				go func(requireTxPoolLock *atomic.Bool) {
+					streams := notificationStreams
+					for {
+						time.Sleep(ethCfg.XLayer.BulkAddTxsWaitTime)
+						if requireTxPoolLock.Load() {
+							continue
+						}
+						streams.Pub(struct{}{})
+					}
+				}(&requireTxPoolLock)
+			}
+		},
+	)
 
 	return tp, nil
 }
