+++ zk/txpool/pool_xlayer.go
@@ -4,6 +4,7 @@ import (
 	"math/big"
 	"strings"
 	"sync"
+	"sync/atomic"
 
 	"github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/types"
@@ -177,6 +178,9 @@ func (p *TxPool) setFreeGasList(freeGasList []ethconfig.FreeGasInfo) {
 }
 
 var notificationStreams *TxpoolNotificationPubSub
+var requireTxPoolLock atomic.Bool
+var notifyOnce sync.Once
+var needNofity bool
 
 type TxpoolNotificationPubSub struct {
 	chans map[uint]chan struct{}
@@ -222,4 +226,10 @@ func (ps *TxpoolNotificationPubSub) remove(id uint) {
 
 func GetNoficationStreams() *TxpoolNotificationPubSub {
 	return notificationStreams
+}
+
+func ArquireTxPoolLock(acquire bool) {
+	if needNofity {
+		requireTxPoolLock.Swap(acquire)
+	}
 }
